backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
boto3==1.34.144
botocore==1.34.144
certifi==2024.7.4
distro==1.9.0
Django==4.2.14
django-cleanup==8.1.0
django-cors-headers==4.4.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
groq==0.9.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
libmagic==1.0
packaging==24.1
pillow==10.4.0
psycopg==3.2.1
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-magic==0.4.27
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==1.26.19
